requirements.txt
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.16
django-allauth==0.42.0
django-cleanup==5.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-currentuser==0.5.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.0
psycopg2-binary==2.8.6
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
